+++ infra/init.sql
@@ -70,5 +70,12 @@ insert into public.issues (id, title, tags) values
     (3, 'Add missing postgrest filters', array['is:open', 'severity:low', 'priority:high']),
     (4, 'Add alias to filters', array['is:closed', 'severity:low', 'priority:medium']);
 
+create or replace function public.list_stored_countries()
+    returns setof countries
+    language sql
+as $function$
+    select * from countries;
+$function$;
+
 alter role authenticator set pgrst.db_plan_enabled to true;
 notify pgrst, 'reload config';