+++ CHANGELOG.md
@@ -2,9 +2,29 @@
 
 
 
+## v0.16.0 (2024-02-27)
+
+### Feature
+
+* feat: Add RPC request builder class for additional filters (#372) ([`0002e8f`](https://github.com/supabase-community/postgrest-py/commit/0002e8f7ec32b6787b44996079b4c2f43fc43717))
+
+
+## v0.15.1 (2024-02-27)
+
+### Chore
+
+* chore(release): bump version to v0.15.1 ([`6d55e49`](https://github.com/supabase-community/postgrest-py/commit/6d55e49b461fd0b52e9267a6b1e47038756bfd7f))
+
+### Fix
+
+* fix: update range to use query parameters instead of headers (#375) ([`eae612c`](https://github.com/supabase-community/postgrest-py/commit/eae612ce0548b392d574e9afc12c11f73e54cf8f))
+
+
 ## v0.15.0 (2024-01-15)
 
 ### Chore
+
+* chore(release): bump version to v0.15.0 ([`0faa8c3`](https://github.com/supabase-community/postgrest-py/commit/0faa8c3f37cb1f360f65dcef075479c297029844))
 
 * chore(deps-dev): bump pytest-asyncio from 0.18.3 to 0.23.3 (#344)